Move component creation from FullCheckoutFlow to CheckoutControllerFactory
COSDK-1117
